pyproject.toml
[project]
name = "patent-system"
version = "0.1.0"
description = "AI-powered patent analysis and drafting system with multi-agent orchestration"
readme = "README.md"
authors = [
    { name = "Hans Koehler", email = "hans@koehler.eu.com" }
]
requires-python = ">=3.13"
dependencies = [
    "nicegui",
    "langgraph",
    "llama-index",
    "dspy",
    "python-docx",
    "pydantic-settings",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[dependency-groups]
dev = [
    "hypothesis",
    "pytest",
]

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"
